feature/command-line-sources-args (#23)
Add command line arguments with yargs.
The only argument so far is a --sources argument, which is a list of
sources based on the source id. This is very useful for testing instead
of worrying about processing all files.

The command line arguments are added to the config in a variable called
`ARGS`, then later get called when pulling the sources. The change for
the different ETL steps is seamless as a result.

Remove a copy-paste import

fixed up a less readable variable name