feature/clustering Added a clustering feature to the tile generator. Tiles will be generated in 3 pieces. The outer zoom is one cluster / city / tile, the middle zoom is regular clustering, and the inner zoom has no clustering, but instead shows all points.
This is a WIP. The outer layer should actually be a geojson layer so
that we can cluster the clusters. The inner layer needs more zoom levels
as it cuts off at 15.

As we are using mbtiles and geojson, we can combine the middle and inner
mbtiles into one big mbtile, then we can combine the outer zoom at one
level to create a geojson of the features at 1 / city.

This was tested with tippecanoe v2.23.0