feature/tree-sources-staging (#24)
* feature/tree-sources-staging

Added a new merge step to the ETL process. The new merge step takes all
the data from the geojson file and moves it into a staging table called
`tree_sources_staging`. This is a way to keep the raw data from the
sources into a database table.

This PR uses ogr2ogr to move data into the database. By default, data is
moved 20k rows at a time. Postgres is all lowercase, and javascript uses
camelcase, so we cannot easily use underscores in names without some
bigger changes. However, that did not seem to be an issue.

feature/tree-sources-staging
Removed unused new library

* feature/tree-sources-staging
- added a new PG_USE_COPY argument. this improves performance by about
2-3x.
- cleaned up command to have shorter strings
- moved PG config to a separate variable

* feature/tree-sources-staging
- update the ogr2ogr command to transform the sql converting camelCase
to snake_case
- convert database table to treedata_staging
- use spawn instead of exec
- truncate the staging table as the first step

Added a new line for a little clean up

* feature/tree-sources-staging
- updated idName to idSourceName as per the rest of the code